fix dev scripts

+++ apps/cli/scripts/installation/uninstall.sh
@@ -1,3 +1,7 @@
 #!/bin/bash
 
-npm uninstall --location=global @withgraphite/graphite-cli
+if command -v volta &>/dev/null; then
+    volta uninstall graphite-motion
+else
+  npm uninstall --location=global @withgraphite/graphite-cli
+fi
